Reword docs on locks to be less strict
Previously the documentation on the 'acquire*' methods said the returned
stage would be immediately complete if the lock was not already held. This
is a fairly strong implementation restriction (notably a synchronous lock
offloading to a threadpool would not meet the requirement) with relatively
limited benefit, so it's better not to guarantee it in the first place.

Some of the lock unit tests may rely on this (now undocumented) behavior.
These are all internal usages which can be changed at a later time